Add marked for markdown rendering post content
Two caveats: not using react-markded, just marked; unsure of the
benefits of using marked with the react wrappers, so I bypassed it.
And I'm adding the .img-fluid class to images after marked runs,
but I'm not using their custom lexer or parser extensions, just
using .replace to add `class="img-fluid"` which could interfere
with any other classes that might be added in the normal course
of rendering markdown.