Add login checks for sensitive API calls, and handle responses
The rails app knows when to return 401 errors for unauthorized
attempts, and the react app knows how to catch these responses
and log/alert/redirect.

It should also do page-load checks on pages like /new and /update
so we aren't getting these 401s. But this should do for now.

Next I can add back in the edit and delete buttons in Post.js.